sk-ollama/requirements.txt
python-dotenv
semantic-kernel
ollama